refactor(DialogTemplate): convert to typescript refactor(ui): import multiple components from ui folder in several files
